requirements.txt
# Telegram Travel Assistant MVP - Minimal Dependencies
# Simplified for serverless deployment on Vercel

# Core AI service (Gemini Vision OCR + Text Generation)
google-generativeai==0.3.2

# Database (Supabase PostgreSQL)
supabase==2.1.0

# Configuration management
python-dotenv==1.0.0

# Image processing (for receipt/ticket OCR)
Pillow==10.0.0
